Fix msvc type warning in Benchmark

+++ gpu/benchmark/Benchmark.hpp
@@ -314,13 +314,14 @@ void postProcessLineData(
     if (sum > NumericType(0)) {
       for (std::size_t k = 0; k < closePoints.size(); ++k) {
         if (weights[k] > NumericType(0)) {
-          closePointsArray.emplace_back(closePoints[k].first, weights[k] / sum);
+          closePointsArray.emplace_back(static_cast<unsigned>(closePoints[k].first), weights[k] / sum);
         }
       }
     } else {
       // Fallback if all weights were discarded
       auto nearestPoint = elementKdTree_->findNearest(diskMesh->nodes[i]);
-      closePointsArray.emplace_back(nearestPoint->first, NumericType(1));
+      closePointsArray.emplace_back(static_cast<unsigned>(nearestPoint->first),
+                                    NumericType(1));
     }
 
     elementsToPoint.push_back(closePointsArray);
